Faster file init and stol for large file size

+++ src/functions.h
@@ -7,8 +7,8 @@ struct Mapping
     char *addr;
     int is_pmem;
     int map_anon; // MAP_ANONYMOUS (not backed by file)
-    int fsize = 0;
-    int buflen = 0;
+    long fsize = 0;
+    long buflen = 0;
 };
 
 void get_bandwidth(int, int, int, Results &);
